Split tmt.hardware into smaller modules
1. files are smaller, easier to navigate
2. separation between constraints and HW requirement parsers
3. because of Artemis slowly transitioning to `tmt.hardware`, we will need
   to include more functionality in tmt.hardware, mostly related to
   flavors - `tmt.hardware.flavors` will be much better than adding to one
   big pile.
